[feature] Added automatic generation of defaul SSH credentials #200
A new SSH keypair will be generated if there does not exist any SSH
key pair on the machine.
The SSH key pair is used to create default access credential which
is used through a default template.
An existing template is updated if it contains configuration to
update the authorized_keys on devices.

Fixes #200